Entity ver1.0
succeed

+++ src/main/java/cp/demo/domain/repository/MenuRepository.java
@@ -3,7 +3,7 @@ package cp.demo.domain.repository;
 import cp.demo.domain.entity.MenuEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuRepository extends JpaRepository<MenuEntity,Integer> {
+public interface MenuRepository extends JpaRepository<MenuEntity, Long>{
 }
